Add files via upload
For ICS 173 final project, data analysis of Maui Nui food insecurity data. 

This code looks through the 200+ categories of calls in the 211 calls summary file (i.e. Rent Payment Assistance, Homeless Shelters) and identifies the strongest signal amongst the categories, as well as the zip codes with the greatest number of calls and the least number of calls for this category.  The identified category has the greatest difference between zip codes. This allows for the identification of the category of greatest need, as well as the geographic areas that need help the most.